+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 
 ## [0.1.3] - 2021-xx-xx
-- TBD
+- Added file checksum to output JSON for canary mode
 
 ## [0.1.2] - 2021-05-11
 - Fixing PGP key bug when JSON output file is not specified